+++ src/components/shared/nav-bar/nav-bar.jsx
@@ -1,16 +1,16 @@
 import Link from 'next/link';
-import RightArrow from 'svg/RightArrow/RightArrow';
-import NoraLogo from 'svg/NoraLogo/NoraLogo';
-import MagGlass from 'svg/MagGlass';
-import styles from './nav.module.scss';
-import Hamburger from 'svg/Hamburger';
-import X from 'svg/X';
+import RightArrow from '@/components/shared/svg/right-arrow/right-arrow';
+import NoraLogo from '@/components/shared/svg/nora-logo/nora-logo';
+import MagGlass from '@/components/shared/svg/magnifying-glass-large';
+import styles from './nav-bar.module.scss';
+import Hamburger from '@/components/shared/svg/hamburger';
+import X from '@/components/shared/svg/close-button';
 import { useState } from 'react';
 import { useRouter } from 'next/router';
 import { useWindowSize } from '@/hooks';
-import Search from '../search/Search';
+import Search from '../search/search-bar';
 
-export default function Nav() {
+export default function NavBar() {
   const [mobileNavView, setMobileNavView] = useState(false);
   const [searchBarView, setSearchBarView] = useState(false);
   const { route } = useRouter();
