+++ src/components/shared/nav-bar/nav-bar.module.scss
