--- src/components/shared/nav/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './nav';
