+++ src/components/shared/search/search-bar.module.scss
