+++ src/components/shared/svg/nora-logo/nora-logo.jsx
@@ -1,8 +1,8 @@
 import Image from 'next/image';
 import ncLogo from 'images/NC.png';
-import styles from './NoraLogo.module.scss';
+import styles from './nora-logo.module.scss';
 
-export default function NavLogo() {
+export default function NoraLogo() {
   return (
     <div className={styles.container}>
       <Image src={ncLogo} alt='main logo' />
