+++ src/components/shared/svg/nora-logo/nora-logo.module.scss
