+++ src/components/shared/svg/right-arrow/right-arrow.module.scss
